+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## Unreleased
+
+### Fixed
+- Bug in authentication handling logic that caused authentication tokens to not automatically renew properly when
+ they expired.
+
+
 ## 0.1.6 – 2019-07-30
 
 ### Fixed